WIP (Move everything to the new graphql_interface macro)
