+++ integration_tests/codegen_fail/fail/interface/argument_double_underscored.rs
@@ -1,15 +1,6 @@
-use juniper::{graphql_interface, GraphQLObject};
-
-#[derive(GraphQLObject)]
-#[graphql(impl = CharacterValue)]
-pub struct ObjA {
-    test: String,
-}
+use juniper::graphql_interface;
 
 #[graphql_interface]
-impl Character for ObjA {}
-
-#[graphql_interface(for = ObjA)]
 trait Character {
     fn id(&self, __num: i32) -> &str {
         "funA"
