+++ integration_tests/codegen_fail/fail/interface/argument_non_input_type.stderr
@@ -1,16 +1,16 @@
 error[E0277]: the trait bound `ObjA: IsInputType<__S>` is not satisfied
-  --> fail/interface/argument_non_input_type.rs:16:1
-   |
-16 | #[graphql_interface(for = ObjA)]
-   | ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ the trait `IsInputType<__S>` is not implemented for `ObjA`
-   |
-   = note: this error originates in the attribute macro `graphql_interface` (in Nightly builds, run with -Z macro-backtrace for more info)
+ --> fail/interface/argument_non_input_type.rs:8:1
+  |
+8 | #[graphql_interface]
+  | ^^^^^^^^^^^^^^^^^^^^ the trait `IsInputType<__S>` is not implemented for `ObjA`
+  |
+  = note: this error originates in the attribute macro `graphql_interface` (in Nightly builds, run with -Z macro-backtrace for more info)
 
 error[E0277]: the trait bound `ObjA: FromInputValue<__S>` is not satisfied
-    --> fail/interface/argument_non_input_type.rs:16:1
+    --> fail/interface/argument_non_input_type.rs:8:1
      |
-16   | #[graphql_interface(for = ObjA)]
-     | ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ the trait `FromInputValue<__S>` is not implemented for `ObjA`
+8    | #[graphql_interface]
+     | ^^^^^^^^^^^^^^^^^^^^ the trait `FromInputValue<__S>` is not implemented for `ObjA`
      |
 note: required by a bound in `Registry::<'r, S>::arg`
     --> $WORKSPACE/juniper/src/executor/mod.rs
@@ -18,11 +18,3 @@ note: required by a bound in `Registry::<'r, S>::arg`
      |         T: GraphQLType<S> + FromInputValue<S>,
      |                             ^^^^^^^^^^^^^^^^^ required by this bound in `Registry::<'r, S>::arg`
      = note: this error originates in the attribute macro `graphql_interface` (in Nightly builds, run with -Z macro-backtrace for more info)
-
-error[E0277]: the trait bound `ObjA: FromInputValue<__S>` is not satisfied
-  --> fail/interface/argument_non_input_type.rs:16:1
-   |
-16 | #[graphql_interface(for = ObjA)]
-   | ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ the trait `FromInputValue<__S>` is not implemented for `ObjA`
-   |
-   = note: this error originates in the attribute macro `graphql_interface` (in Nightly builds, run with -Z macro-backtrace for more info)
