+++ integration_tests/codegen_fail/fail/interface/argument_wrong_default_array.rs
@@ -1,21 +1,8 @@
-use juniper::{graphql_interface, GraphQLObject};
-
-#[derive(GraphQLObject)]
-#[graphql(impl = CharacterValue)]
-pub struct ObjA {
-    test: String,
-}
+use juniper::graphql_interface;
 
 #[graphql_interface]
-impl Character for ObjA {}
-
-#[graphql_interface(for = ObjA)]
 trait Character {
-    fn wrong(
-        &self,
-        #[graphql(default = [true, false, false])]
-        input: [bool; 2],
-    ) -> bool {
+    fn wrong(&self, #[graphql(default = [true, false, false])] input: [bool; 2]) -> bool {
         input[0]
     }
 }
