+++ integration_tests/codegen_fail/fail/interface/argument_wrong_default_array.stderr
@@ -1,11 +1,11 @@
 error[E0277]: the trait bound `[bool; 2]: From<[bool; 3]>` is not satisfied
-  --> fail/interface/argument_wrong_default_array.rs:12:1
-   |
-12 | #[graphql_interface(for = ObjA)]
-   | ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ the trait `From<[bool; 3]>` is not implemented for `[bool; 2]`
-   |
-   = help: the following implementations were found:
-             <[T; LANES] as From<Simd<T, LANES>>>
-             <[bool; LANES] as From<Mask<T, LANES>>>
-   = note: required because of the requirements on the impl of `Into<[bool; 2]>` for `[bool; 3]`
-   = note: this error originates in the attribute macro `graphql_interface` (in Nightly builds, run with -Z macro-backtrace for more info)
+ --> fail/interface/argument_wrong_default_array.rs:3:1
+  |
+3 | #[graphql_interface]
+  | ^^^^^^^^^^^^^^^^^^^^ the trait `From<[bool; 3]>` is not implemented for `[bool; 2]`
+  |
+  = help: the following implementations were found:
+            <[T; LANES] as From<Simd<T, LANES>>>
+            <[bool; LANES] as From<Mask<T, LANES>>>
+  = note: required because of the requirements on the impl of `Into<[bool; 2]>` for `[bool; 3]`
+  = note: this error originates in the attribute macro `graphql_interface` (in Nightly builds, run with -Z macro-backtrace for more info)
