--- integration_tests/codegen_fail/fail/interface/downcast_method_wrong_return_type.stderr
@@ -1,19 +1,0 @@
-error: GraphQL interface expects trait method return type to be `Option<&ImplementerType>` only
-  --> fail/interface/downcast_method_wrong_return_type.rs:10:40
-   |
-10 |     fn a(&self, ctx: &(), rand: u8) -> &Human {
-   |                                        ^^^^^^
-   |
-   = note: https://spec.graphql.org/June2018/#sec-Interfaces
-
-error[E0412]: cannot find type `CharacterValue` in this scope
-  --> fail/interface/downcast_method_wrong_return_type.rs:16:18
-   |
-16 | #[graphql(impl = CharacterValue)]
-   |                  ^^^^^^^^^^^^^^ not found in this scope
-
-error[E0405]: cannot find trait `Character` in this scope
-  --> fail/interface/downcast_method_wrong_return_type.rs:22:6
-   |
-22 | impl Character for Human {}
-   |      ^^^^^^^^^ not found in this scope
