+++ integration_tests/codegen_fail/fail/interface/field_double_underscored.stderr
@@ -1,19 +1,7 @@
 error: All types and directives defined within a schema must not have a name which begins with `__` (two underscores), as this is used exclusively by GraphQL’s introspection system.
-  --> fail/interface/field_double_underscored.rs:14:8
-   |
-14 |     fn __id(&self) -> &str {
-   |        ^^^^
-   |
-   = note: https://spec.graphql.org/June2018/#sec-Schema
-
-error[E0412]: cannot find type `CharacterValue` in this scope
- --> fail/interface/field_double_underscored.rs:4:18
+ --> fail/interface/field_double_underscored.rs:5:8
   |
-4 | #[graphql(impl = CharacterValue)]
-  |                  ^^^^^^^^^^^^^^ not found in this scope
-
-error[E0405]: cannot find trait `Character` in this scope
-  --> fail/interface/field_double_underscored.rs:10:6
-   |
-10 | impl Character for ObjA {}
-   |      ^^^^^^^^^ not found in this scope
+5 |     fn __id(&self) -> &str {
+  |        ^^^^
+  |
+  = note: https://spec.graphql.org/June2018/#sec-Schema
