+++ integration_tests/codegen_fail/fail/interface/field_non_output_return_type.rs
@@ -1,24 +1,13 @@
-use juniper::{graphql_interface, GraphQLInputObject, GraphQLObject};
+use juniper::{graphql_interface, GraphQLInputObject};
 
 #[derive(GraphQLInputObject)]
 pub struct ObjB {
     id: i32,
 }
 
-#[derive(GraphQLObject)]
-#[graphql(impl = CharacterValue)]
-pub struct ObjA {
-    test: String,
-}
-
 #[graphql_interface]
-impl Character for ObjA {}
-
-#[graphql_interface(for = ObjA)]
 trait Character {
-    fn id(&self) -> ObjB {
-        ObjB { id: 34 }
-    }
+    fn id(&self) -> ObjB;
 }
 
 fn main() {}
