+++ integration_tests/codegen_fail/fail/interface/fields_duplicate.rs
@@ -1,24 +1,11 @@
-use juniper::{graphql_interface, GraphQLObject};
-
-#[derive(GraphQLObject)]
-#[graphql(impl = CharacterValue)]
-pub struct ObjA {
-    test: String,
-}
+use juniper::graphql_interface;
 
 #[graphql_interface]
-impl Character for ObjA {}
-
-#[graphql_interface(for = ObjA)]
 trait Character {
-    fn id(&self) -> &str {
-        "funA"
-    }
+    fn id(&self) -> &str;
 
     #[graphql(name = "id")]
-    fn id2(&self) -> &str {
-        "funB"
-    }
+    fn id2(&self) -> &str;
 }
 
 fn main() {}
