--- integration_tests/codegen_fail/fail/interface/implementer_non_object_type.stderr
@@ -1,15 +1,0 @@
-error[E0277]: the trait bound `ObjA: GraphQLObject<__S>` is not satisfied
-  --> fail/interface/implementer_non_object_type.rs:15:1
-   |
-15 | #[graphql_interface(for = ObjA)]
-   | ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ the trait `GraphQLObject<__S>` is not implemented for `ObjA`
-   |
-   = note: this error originates in the attribute macro `graphql_interface` (in Nightly builds, run with -Z macro-backtrace for more info)
-
-error[E0277]: the trait bound `ObjA: IsOutputType<__S>` is not satisfied
-  --> fail/interface/implementer_non_object_type.rs:15:1
-   |
-15 | #[graphql_interface(for = ObjA)]
-   | ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ the trait `IsOutputType<__S>` is not implemented for `ObjA`
-   |
-   = note: this error originates in the attribute macro `graphql_interface` (in Nightly builds, run with -Z macro-backtrace for more info)
