+++ integration_tests/codegen_fail/fail/interface/implementers_duplicate_pretty.rs
@@ -3,14 +3,7 @@ use juniper::{graphql_interface, GraphQLObject};
 #[derive(GraphQLObject)]
 #[graphql(impl = CharacterValue)]
 pub struct ObjA {
-    test: String,
-}
-
-#[graphql_interface]
-impl Character for ObjA {
-    fn id(&self) -> &str {
-        "funA"
-    }
+    id: String,
 }
 
 #[graphql_interface(for = [ObjA, ObjA])]
