+++ integration_tests/codegen_fail/fail/interface/implementers_duplicate_ugly.rs
@@ -3,17 +3,10 @@ use juniper::{graphql_interface, GraphQLObject};
 #[derive(GraphQLObject)]
 #[graphql(impl = CharacterValue)]
 pub struct ObjA {
-    test: String,
+    id: String,
 }
 
 type ObjAlias = ObjA;
-
-#[graphql_interface]
-impl Character for ObjA {
-    fn id(&self) -> &str {
-        "funA"
-    }
-}
 
 #[graphql_interface(for = [ObjA, ObjAlias])]
 trait Character {
