+++ integration_tests/codegen_fail/fail/interface/implementers_duplicate_ugly.stderr
@@ -1,7 +1,7 @@
-error[E0119]: conflicting implementations of trait `<CharacterValue as juniper::GraphQLInterface<__S>>::mark::_::{closure#0}::MutuallyExclusive` for type `ObjA`
-  --> $DIR/implementers_duplicate_ugly.rs:18:1
+error[E0119]: conflicting implementations of trait `<CharacterValueEnum<ObjA, ObjA> as juniper::GraphQLInterface<__S>>::mark::_::{closure#0}::MutuallyExclusive` for type `ObjA`
+  --> fail/interface/implementers_duplicate_ugly.rs:11:1
    |
-18 | #[graphql_interface(for = [ObjA, ObjAlias])]
+11 | #[graphql_interface(for = [ObjA, ObjAlias])]
    | ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
    | |
    | first implementation here
@@ -9,13 +9,13 @@ error[E0119]: conflicting implementations of trait `<CharacterValue as juniper::
    |
    = note: this error originates in the macro `::juniper::sa::assert_type_ne_all` (in Nightly builds, run with -Z macro-backtrace for more info)
 
-error[E0119]: conflicting implementations of trait `std::convert::From<ObjA>` for type `CharacterValue`
-  --> $DIR/implementers_duplicate_ugly.rs:18:1
+error[E0119]: conflicting implementations of trait `std::convert::From<ObjA>` for type `CharacterValueEnum<ObjA, ObjA>`
+  --> fail/interface/implementers_duplicate_ugly.rs:11:1
    |
-18 | #[graphql_interface(for = [ObjA, ObjAlias])]
+11 | #[graphql_interface(for = [ObjA, ObjAlias])]
    | ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
    | |
    | first implementation here
-   | conflicting implementation for `CharacterValue`
+   | conflicting implementation for `CharacterValueEnum<ObjA, ObjA>`
    |
    = note: this error originates in the attribute macro `graphql_interface` (in Nightly builds, run with -Z macro-backtrace for more info)
