+++ integration_tests/codegen_fail/fail/interface/name_double_underscored.rs
@@ -2,9 +2,7 @@ use juniper::graphql_interface;
 
 #[graphql_interface]
 trait __Character {
-    fn id(&self) -> &str {
-        "funA"
-    }
+    fn id(&self) -> &str;
 }
 
 fn main() {}
