+++ integration_tests/codegen_fail/fail/interface/no_fields.stderr
@@ -1,19 +1,7 @@
 error: GraphQL interface must have at least one field
-  --> fail/interface/no_fields.rs:13:1
-   |
-13 | trait Character {}
-   | ^^^^^^^^^^^^^^^^^^
-   |
-   = note: https://spec.graphql.org/June2018/#sec-Interfaces
-
-error[E0412]: cannot find type `CharacterValue` in this scope
- --> fail/interface/no_fields.rs:4:18
+ --> fail/interface/no_fields.rs:4:1
   |
-4 | #[graphql(impl = CharacterValue)]
-  |                  ^^^^^^^^^^^^^^ not found in this scope
-
-error[E0405]: cannot find trait `Character` in this scope
-  --> fail/interface/no_fields.rs:10:6
-   |
-10 | impl Character for ObjA {}
-   |      ^^^^^^^^^ not found in this scope
+4 | trait Character {}
+  | ^^^^^^^^^^^^^^^^^^
+  |
+  = note: https://spec.graphql.org/June2018/#sec-Interfaces
