+++ integration_tests/codegen_fail/fail/interface/wrong_item_enum.rs
