+++ juniper_codegen/src/graphql_interface/mod.rs
@@ -3,21 +3,19 @@
 //! [1]: https://spec.graphql.org/June2018/#sec-Interfaces
 
 pub mod attr;
-pub mod new;
 
-use std::{
-    collections::{HashMap, HashSet},
-    convert::TryInto as _,
-};
+use std::{collections::HashSet, convert::TryInto as _};
 
 use proc_macro2::TokenStream;
-use quote::{format_ident, quote, ToTokens, TokenStreamExt as _};
+use quote::{format_ident, quote, ToTokens};
 use syn::{
     ext::IdentExt as _,
     parse::{Parse, ParseStream},
     parse_quote,
+    punctuated::Punctuated,
     spanned::Spanned as _,
     token,
+    visit::Visit,
 };
 
 use crate::{
@@ -53,32 +51,21 @@ struct TraitAttr {
     /// [2]: https://spec.graphql.org/June2018/#sec-Descriptions
     description: Option<SpanContainer<String>>,
 
-    /// Explicitly specified identifier of the enum Rust type behind the trait,
-    /// being an actual implementation of a [GraphQL interface][1] type.
+    /// Explicitly specified identifier of the type alias of Rust enum type
+    /// behind the trait, being an actual implementation of a
+    /// [GraphQL interface][1] type.
     ///
     /// If [`None`], then `{trait_name}Value` identifier will be used.
     ///
     /// [1]: https://spec.graphql.org/June2018/#sec-Interfaces
     r#enum: Option<SpanContainer<syn::Ident>>,
 
-    /// Explicitly specified identifier of the Rust type alias of the
-    /// [trait object][2], being an actual implementation of a
-    /// [GraphQL interface][1] type.
-    ///
-    /// Effectively makes code generation to use a [trait object][2] as a
-    /// [GraphQL interface][1] type rather than an enum. If [`None`], then enum
-    /// is used by default.
-    ///
-    /// [1]: https://spec.graphql.org/June2018/#sec-Interfaces
-    /// [2]: https://doc.rust-lang.org/reference/types/trait-object.html
-    r#dyn: Option<SpanContainer<syn::Ident>>,
-
     /// Explicitly specified Rust types of [GraphQL objects][2] implementing
     /// this [GraphQL interface][1] type.
     ///
     /// [1]: https://spec.graphql.org/June2018/#sec-Interfaces
     /// [2]: https://spec.graphql.org/June2018/#sec-Objects
-    implementers: HashSet<SpanContainer<syn::Type>>,
+    implementers: HashSet<SpanContainer<syn::TypePath>>,
 
     /// Explicitly specified type of [`Context`] to use for resolving this
     /// [GraphQL interface][1] type with.
@@ -109,22 +96,6 @@ struct TraitAttr {
     /// If [`None`], then trait will be transformed into [`async_trait`] only if
     /// it contains async methods.
     asyncness: Option<SpanContainer<syn::Ident>>,
-
-    /// Explicitly specified external downcasting functions for
-    /// [GraphQL interface][1] implementers.
-    ///
-    /// If [`None`], then macro will downcast to the implementers via enum
-    /// dispatch or dynamic dispatch (if the one is chosen). That's why
-    /// specifying an external resolver function has sense, when some custom
-    /// [interface][1] implementer resolving logic is involved.
-    ///
-    /// Once the downcasting function is specified for some [GraphQL object][2]
-    /// implementer type, it cannot be downcast another such function or trait
-    /// method marked with a [`MethodMeta::downcast`] marker.
-    ///
-    /// [1]: https://spec.graphql.org/June2018/#sec-Interfaces
-    /// [2]: https://spec.graphql.org/June2018/#sec-Objects
-    external_downcasts: HashMap<syn::Type, SpanContainer<syn::ExprPath>>,
 
     /// Explicitly specified [`RenameRule`] for all fields of this
     /// [GraphQL interface][1] type.
@@ -184,7 +155,7 @@ impl Parse for TraitAttr {
                 "for" | "implementers" => {
                     input.parse::<token::Eq>()?;
                     for impler in input.parse_maybe_wrapped_and_punctuated::<
-                        syn::Type, token::Bracket, token::Comma,
+                        syn::TypePath, token::Bracket, token::Comma,
                     >()? {
                         let impler_span = impler.span();
                         out
@@ -192,13 +163,6 @@ impl Parse for TraitAttr {
                             .replace(SpanContainer::new(ident.span(), Some(impler_span), impler))
                             .none_or_else(|_| err::dup_arg(impler_span))?;
                     }
-                }
-                "dyn" => {
-                    input.parse::<token::Eq>()?;
-                    let alias = input.parse::<syn::Ident>()?;
-                    out.r#dyn
-                        .replace(SpanContainer::new(ident.span(), Some(alias.span()), alias))
-                        .none_or_else(|_| err::dup_arg(&ident))?
                 }
                 "enum" => {
                     input.parse::<token::Eq>()?;
@@ -212,16 +176,6 @@ impl Parse for TraitAttr {
                     out.asyncness
                         .replace(SpanContainer::new(span, Some(span), ident))
                         .none_or_else(|_| err::dup_arg(span))?;
-                }
-                "on" => {
-                    let ty = input.parse::<syn::Type>()?;
-                    input.parse::<token::Eq>()?;
-                    let dwncst = input.parse::<syn::ExprPath>()?;
-                    let dwncst_spanned = SpanContainer::new(ident.span(), Some(ty.span()), dwncst);
-                    let dwncst_span = dwncst_spanned.span_joined();
-                    out.external_downcasts
-                        .insert(ty, dwncst_spanned)
-                        .none_or_else(|_| err::dup_arg(dwncst_span))?
                 }
                 "rename_all" => {
                     input.parse::<token::Eq>()?;
@@ -257,12 +211,8 @@ impl TraitAttr {
             context: try_merge_opt!(context: self, another),
             scalar: try_merge_opt!(scalar: self, another),
             implementers: try_merge_hashset!(implementers: self, another => span_joined),
-            r#dyn: try_merge_opt!(r#dyn: self, another),
             r#enum: try_merge_opt!(r#enum: self, another),
             asyncness: try_merge_opt!(asyncness: self, another),
-            external_downcasts: try_merge_hashmap!(
-                external_downcasts: self, another => span_joined
-            ),
             rename_fields: try_merge_opt!(rename_fields: self, another),
             is_internal: self.is_internal || another.is_internal,
         })
@@ -275,15 +225,6 @@ impl TraitAttr {
             .map(|attr| attr.parse_args())
             .try_fold(Self::default(), |prev, curr| prev.try_merge(curr?))?;
 
-        if let Some(as_dyn) = &attr.r#dyn {
-            if attr.r#enum.is_some() {
-                return Err(syn::Error::new(
-                    as_dyn.span(),
-                    "`dyn` attribute argument is not composable with `enum` attribute argument",
-                ));
-            }
-        }
-
         if attr.description.is_none() {
             attr.description = get_doc_comment(attrs);
         }
@@ -292,109 +233,32 @@ impl TraitAttr {
     }
 }
 
-/// Available arguments behind `#[graphql_interface]` attribute placed on a
-/// trait implementation block, when generating code for [GraphQL interface][1]
-/// type.
-///
-/// [1]: https://spec.graphql.org/June2018/#sec-Interfaces
-#[derive(Debug, Default)]
-struct ImplAttr {
-    /// Explicitly specified type (or type parameter with its bounds) of
-    /// [`ScalarValue`] to implementing the [GraphQL interface][1] type with.
-    ///
-    /// If absent, then generated code will be generic over any [`ScalarValue`]
-    /// type, which, in turn, requires all [interface][1] implementers to be
-    /// generic over any [`ScalarValue`] type too. That's why this type should
-    /// be specified only if the implementer itself implements [`GraphQLType`]
-    /// in a non-generic way over [`ScalarValue`] type.
-    ///
-    /// [`GraphQLType`]: juniper::GraphQLType
-    /// [`ScalarValue`]: juniper::ScalarValue
-    /// [1]: https://spec.graphql.org/June2018/#sec-Interfaces
-    scalar: Option<SpanContainer<scalar::AttrValue>>,
-
-    /// Explicitly specified marker indicating that the trait implementation
-    /// block should be transformed with applying [`async_trait`].
-    ///
-    /// If absent, then trait will be transformed with applying [`async_trait`]
-    /// only if it contains async methods.
-    ///
-    /// This marker is especially useful when Rust trait contains async default
-    /// methods, while the implementation block doesn't.
-    asyncness: Option<SpanContainer<syn::Ident>>,
-
-    /// Explicitly specified marker indicating that the implemented
-    /// [GraphQL interface][1] type is represented as a [trait object][2] in
-    /// Rust type system rather then an enum (default mode, when the marker is
-    /// absent).
-    ///
-    /// [2]: https://doc.rust-lang.org/reference/types/trait-object.html
-    r#dyn: Option<SpanContainer<syn::Ident>>,
-}
-
-impl Parse for ImplAttr {
-    fn parse(input: ParseStream<'_>) -> syn::Result<Self> {
-        let mut out = Self::default();
-        while !input.is_empty() {
-            let ident = input.parse_any_ident()?;
-            match ident.to_string().as_str() {
-                "scalar" | "Scalar" | "ScalarValue" => {
-                    input.parse::<token::Eq>()?;
-                    let scl = input.parse::<scalar::AttrValue>()?;
-                    out.scalar
-                        .replace(SpanContainer::new(ident.span(), Some(scl.span()), scl))
-                        .none_or_else(|_| err::dup_arg(&ident))?
-                }
-                "dyn" => {
-                    let span = ident.span();
-                    out.r#dyn
-                        .replace(SpanContainer::new(span, Some(span), ident))
-                        .none_or_else(|_| err::dup_arg(span))?;
-                }
-                "async" => {
-                    let span = ident.span();
-                    out.asyncness
-                        .replace(SpanContainer::new(span, Some(span), ident))
-                        .none_or_else(|_| err::dup_arg(span))?;
-                }
-                name => {
-                    return Err(err::unknown_arg(&ident, name));
-                }
-            }
-            input.try_parse::<token::Comma>()?;
-        }
-        Ok(out)
-    }
-}
-
-impl ImplAttr {
-    /// Tries to merge two [`ImplAttr`]s into a single one, reporting about
-    /// duplicates, if any.
-    fn try_merge(self, mut another: Self) -> syn::Result<Self> {
-        Ok(Self {
-            scalar: try_merge_opt!(scalar: self, another),
-            r#dyn: try_merge_opt!(r#dyn: self, another),
-            asyncness: try_merge_opt!(asyncness: self, another),
-        })
-    }
-
-    /// Parses [`ImplAttr`] from the given multiple `name`d [`syn::Attribute`]s
-    /// placed on a trait implementation block.
-    pub fn from_attrs(name: &str, attrs: &[syn::Attribute]) -> syn::Result<Self> {
-        filter_attrs(name, attrs)
-            .map(|attr| attr.parse_args())
-            .try_fold(Self::default(), |prev, curr| prev.try_merge(curr?))
-    }
-}
-
 /// Definition of [GraphQL interface][1] for code generation.
 ///
 /// [1]: https://spec.graphql.org/June2018/#sec-Interfaces
 struct Definition {
-    /// Rust type that this [GraphQL interface][1] is represented with.
+    /// [`syn::Generics`] of the trait describing the [GraphQL interface][1].
     ///
     /// [1]: https://spec.graphql.org/June2018/#sec-Interfaces
-    ty: Type,
+    trait_generics: syn::Generics,
+
+    /// [`syn::Visibility`] of the trait describing the [GraphQL interface][1].
+    ///
+    /// [1]: https://spec.graphql.org/June2018/#sec-Interfaces
+    vis: syn::Visibility,
+
+    /// Name of the generic enum describing all [`implementers`]. It's generic
+    /// to derive [`Clone`], [`Copy`] and [`Debug`] on it.
+    ///
+    /// [`implementers`]: Self::implementers
+    /// [`Debug`]: std::fmt::Debug
+    enum_ident: syn::Ident,
+
+    /// Name of the type alias for [`enum_ident`] with [`implementers`].
+    ///
+    /// [`enum_ident`]: Self::enum_ident
+    /// [`implementers`]: Self::implementers
+    enum_alias_ident: syn::Ident,
 
     /// Name of this [GraphQL interface][1] in GraphQL schema.
     ///
@@ -431,22 +295,142 @@ struct Definition {
     /// Defined [`Implementer`]s of this [GraphQL interface][1].
     ///
     /// [1]: https://spec.graphql.org/June2018/#sec-Interfaces
-    implementers: Vec<Implementer>,
+    implementers: Vec<syn::TypePath>,
 }
 
 impl ToTokens for Definition {
     fn to_tokens(&self, into: &mut TokenStream) {
-        self.ty.to_token_stream().to_tokens(into);
+        self.generate_enum_tokens().to_tokens(into);
         self.impl_graphql_interface_tokens().to_tokens(into);
         self.impl_output_type_tokens().to_tokens(into);
         self.impl_graphql_type_tokens().to_tokens(into);
         self.impl_graphql_value_tokens().to_tokens(into);
         self.impl_graphql_value_async_tokens().to_tokens(into);
         self.impl_traits_for_reflection_tokens().to_tokens(into);
+        self.impl_field_meta_tokens().to_tokens(into);
+        self.impl_field_tokens().to_tokens(into);
+        self.impl_async_field_tokens().to_tokens(into);
     }
 }
 
 impl Definition {
+    /// Generates enum describing all [`implementers`].
+    ///
+    /// [`implementers`]: Self::implementers
+    #[must_use]
+    fn generate_enum_tokens(&self) -> TokenStream {
+        let vis = &self.vis;
+        let enum_ident = &self.enum_ident;
+        let alias_ident = &self.enum_alias_ident;
+
+        let variant_gens_pars = self
+            .implementers
+            .iter()
+            .enumerate()
+            .map::<syn::GenericParam, _>(|(id, _)| {
+                let par = format_ident!("__I{}", id);
+                parse_quote!( #par )
+            });
+
+        let variants_idents = self
+            .implementers
+            .iter()
+            .filter_map(|ty| ty.path.segments.last().map(|seg| &seg.ident));
+
+        let trait_gens = &self.trait_generics;
+        let (trait_impl_gens, trait_ty_gens, trait_where_clause) =
+            self.trait_generics.split_for_impl();
+
+        let (trait_gens_lifetimes, trait_gens_tys) = trait_gens
+            .params
+            .clone()
+            .into_iter()
+            .partition::<Punctuated<_, _>, _>(|par| {
+            matches!(par, syn::GenericParam::Lifetime(_))
+        });
+
+        let enum_gens = {
+            let mut enum_gens = trait_gens.clone();
+            enum_gens.params = trait_gens_lifetimes.clone();
+            enum_gens.params.extend(variant_gens_pars.clone());
+            enum_gens.params.extend(trait_gens_tys.clone());
+            enum_gens
+        };
+
+        let enum_alias_gens = {
+            let mut enum_alias_gens = trait_gens.clone();
+            enum_alias_gens.move_bounds_to_where_clause();
+            enum_alias_gens
+        };
+
+        let enum_to_alias_gens = {
+            trait_gens_lifetimes
+                .into_iter()
+                .map(|par| match par {
+                    syn::GenericParam::Lifetime(def) => {
+                        let lifetime = &def.lifetime;
+                        quote! { #lifetime }
+                    }
+                    rest => quote! { #rest },
+                })
+                .chain(self.implementers.iter().map(ToTokens::to_token_stream))
+                .chain(trait_gens_tys.into_iter().map(|par| match par {
+                    syn::GenericParam::Type(ty) => {
+                        let par_ident = &ty.ident;
+                        quote! { #par_ident }
+                    }
+                    rest => quote! { #rest },
+                }))
+        };
+
+        let phantom_variant = self.has_phantom_variant().then(|| {
+            let phantom_params = trait_gens.params.iter().filter_map(|p| {
+                let ty = match p {
+                    syn::GenericParam::Type(ty) => {
+                        let ident = &ty.ident;
+                        quote! { #ident }
+                    }
+                    syn::GenericParam::Lifetime(lt) => {
+                        let lifetime = &lt.lifetime;
+                        quote! { &#lifetime () }
+                    }
+                    syn::GenericParam::Const(_) => return None,
+                };
+                Some(quote! {
+                    ::std::marker::PhantomData<::std::sync::atomic::AtomicPtr<Box<#ty>>>
+                })
+            });
+            quote! { __Phantom(#(#phantom_params),*) }
+        });
+
+        let from_impls = self.implementers.iter().zip(variants_idents.clone()).map(
+            |(ty, ident)| {
+                quote! {
+                    impl#trait_impl_gens ::std::convert::From<#ty> for #alias_ident#trait_ty_gens
+                        #trait_where_clause
+                    {
+                        fn from(v: #ty) -> Self {
+                            Self::#ident(v)
+                        }
+                    }
+                }
+            },
+        );
+
+        quote! {
+            #[derive(Clone, Copy, Debug)]
+            #vis enum #enum_ident#enum_gens {
+                #(#variants_idents(#variant_gens_pars),)*
+                #phantom_variant
+            }
+
+            #vis type #alias_ident#enum_alias_gens =
+                #enum_ident<#(#enum_to_alias_gens),*>;
+
+            #(#from_impls)*
+        }
+    }
+
     /// Returns generated code implementing [`GraphQLInterface`] trait for this
     /// [GraphQL interface][1].
     ///
@@ -454,19 +438,23 @@ impl Definition {
     /// [1]: https://spec.graphql.org/June2018/#sec-Interfaces
     #[must_use]
     fn impl_graphql_interface_tokens(&self) -> TokenStream {
+        let ty = &self.enum_alias_ident;
         let scalar = &self.scalar;
 
-        let (impl_generics, where_clause) = self.ty.impl_generics(false);
-        let ty = self.ty.ty_tokens();
+        let gens = self.impl_generics(false);
+        let (impl_generics, _, where_clause) = gens.split_for_impl();
+        let (_, ty_generics, _) = self.trait_generics.split_for_impl();
 
-        let impler_tys: Vec<_> = self.implementers.iter().map(|impler| &impler.ty).collect();
+        let impler_tys = &self.implementers;
         let all_implers_unique = (impler_tys.len() > 1).then(|| {
             quote! { ::juniper::sa::assert_type_ne_all!(#( #impler_tys ),*); }
         });
 
         quote! {
             #[automatically_derived]
-            impl#impl_generics ::juniper::marker::GraphQLInterface<#scalar> for #ty #where_clause
+            impl#impl_generics ::juniper::marker::GraphQLInterface<#scalar> for
+                #ty#ty_generics
+                #where_clause
             {
                 fn mark() {
                     #all_implers_unique
@@ -483,21 +471,25 @@ impl Definition {
     /// [1]: https://spec.graphql.org/June2018/#sec-Interfaces
     #[must_use]
     fn impl_output_type_tokens(&self) -> TokenStream {
+        let ty = &self.enum_alias_ident;
         let scalar = &self.scalar;
 
-        let (impl_generics, where_clause) = self.ty.impl_generics(false);
-        let ty = self.ty.ty_tokens();
+        let generics = self.impl_generics(false);
+        let (impl_generics, _, where_clause) = generics.split_for_impl();
+        let (_, ty_generics, _) = self.trait_generics.split_for_impl();
 
         let fields_marks = self
             .fields
             .iter()
             .map(|f| f.method_mark_tokens(false, scalar));
 
-        let impler_tys = self.implementers.iter().map(|impler| &impler.ty);
+        let impler_tys = self.implementers.iter();
 
         quote! {
             #[automatically_derived]
-            impl#impl_generics ::juniper::marker::IsOutputType<#scalar> for #ty #where_clause
+            impl#impl_generics ::juniper::marker::IsOutputType<#scalar> for
+                #ty#ty_generics
+                #where_clause
             {
                 fn mark() {
                     #( #fields_marks )*
@@ -514,10 +506,12 @@ impl Definition {
     /// [1]: https://spec.graphql.org/June2018/#sec-Interfaces
     #[must_use]
     fn impl_graphql_type_tokens(&self) -> TokenStream {
+        let ty = &self.enum_alias_ident;
         let scalar = &self.scalar;
 
-        let (impl_generics, where_clause) = self.ty.impl_generics(false);
-        let ty = self.ty.ty_tokens();
+        let generics = self.impl_generics(false);
+        let (impl_generics, _, where_clause) = generics.split_for_impl();
+        let (_, ty_generics, _) = self.trait_generics.split_for_impl();
 
         let name = &self.name;
         let description = self
@@ -526,7 +520,7 @@ impl Definition {
             .map(|desc| quote! { .description(#desc) });
 
         // Sorting is required to preserve/guarantee the order of implementers registered in schema.
-        let mut impler_tys: Vec<_> = self.implementers.iter().map(|impler| &impler.ty).collect();
+        let mut impler_tys = self.implementers.clone();
         impler_tys.sort_unstable_by(|a, b| {
             let (a, b) = (quote!(#a).to_string(), quote!(#b).to_string());
             a.cmp(&b)
@@ -536,7 +530,9 @@ impl Definition {
 
         quote! {
             #[automatically_derived]
-            impl#impl_generics ::juniper::GraphQLType<#scalar> for #ty #where_clause
+            impl#impl_generics ::juniper::GraphQLType<#scalar>
+                for #ty#ty_generics
+                #where_clause
             {
                 fn name(_ : &Self::TypeInfo) -> Option<&'static str> {
                     Some(#name)
@@ -554,7 +550,7 @@ impl Definition {
                     let fields = [
                         #( #fields_meta, )*
                     ];
-                    registry.build_interface_type::<#ty>(info, &fields)
+                    registry.build_interface_type::<#ty#ty_generics>(info, &fields)
                         #description
                         .into_meta()
                 }
@@ -569,18 +565,30 @@ impl Definition {
     /// [1]: https://spec.graphql.org/June2018/#sec-Interfaces
     #[must_use]
     fn impl_graphql_value_tokens(&self) -> TokenStream {
+        let ty = &self.enum_alias_ident;
+        let trait_name = &self.name;
         let scalar = &self.scalar;
         let context = &self.context;
 
-        let (impl_generics, where_clause) = self.ty.impl_generics(false);
-        let ty = self.ty.ty_tokens();
-        let ty_name = ty.to_string();
-        let trait_ty = self.ty.trait_ty();
+        let generics = self.impl_generics(false);
+        let (impl_generics, _, where_clause) = generics.split_for_impl();
+        let (_, ty_generics, _) = self.trait_generics.split_for_impl();
 
-        let fields_resolvers = self
-            .fields
-            .iter()
-            .filter_map(|f| f.method_resolve_field_tokens(scalar, Some(&trait_ty)));
+        let fields_resolvers = self.fields.iter().filter_map(|f| {
+            if f.is_async {
+                return None;
+            }
+
+            let name = &f.name;
+            Some(quote! {
+                #name => {
+                    ::juniper::macros::reflection::Field::<
+                        #scalar,
+                        { ::juniper::macros::reflection::fnv1a128(#name) }
+                    >::call(self, info, args, executor)
+                }
+            })
+        });
         let async_fields_err = {
             let names = self
                 .fields
@@ -589,29 +597,22 @@ impl Definition {
                 .collect::<Vec<_>>();
             (!names.is_empty()).then(|| {
                 field::Definition::method_resolve_field_err_async_field_tokens(
-                    &names, scalar, &ty_name,
+                    &names, scalar, trait_name,
                 )
             })
         };
         let no_field_err =
-            field::Definition::method_resolve_field_err_no_field_tokens(scalar, &ty_name);
+            field::Definition::method_resolve_field_err_no_field_tokens(scalar, trait_name);
 
-        let custom_downcast_checks = self
-            .implementers
-            .iter()
-            .filter_map(|i| i.method_concrete_type_name_tokens(&trait_ty));
-        let regular_downcast_check = self.ty.method_concrete_type_name_tokens();
+        let downcast_check = self.method_concrete_type_name_tokens();
 
-        let custom_downcasts = self
-            .implementers
-            .iter()
-            .filter_map(|i| i.method_resolve_into_type_tokens(&trait_ty));
-        let regular_downcast = self.ty.method_resolve_into_type_tokens();
+        let downcast = self.method_resolve_into_type_tokens();
 
         quote! {
             #[allow(deprecated)]
             #[automatically_derived]
-            impl#impl_generics ::juniper::GraphQLValue<#scalar> for #ty #where_clause
+            impl#impl_generics ::juniper::GraphQLValue<#scalar> for #ty#ty_generics
+                #where_clause
             {
                 type Context = #context;
                 type TypeInfo = ();
@@ -639,8 +640,7 @@ impl Definition {
                     context: &Self::Context,
                     info: &Self::TypeInfo,
                 ) -> String {
-                    #( #custom_downcast_checks )*
-                    #regular_downcast_check
+                    #downcast_check
                 }
 
                 fn resolve_into_type(
@@ -650,8 +650,7 @@ impl Definition {
                     _: Option<&[::juniper::Selection<#scalar>]>,
                     executor: &::juniper::Executor<Self::Context, #scalar>,
                 ) -> ::juniper::ExecutionResult<#scalar> {
-                    #( #custom_downcasts )*
-                    #regular_downcast
+                    #downcast
                 }
             }
         }
@@ -664,30 +663,35 @@ impl Definition {
     /// [1]: https://spec.graphql.org/June2018/#sec-Interfaces
     #[must_use]
     fn impl_graphql_value_async_tokens(&self) -> TokenStream {
+        let ty = &self.enum_alias_ident;
+        let trait_name = &self.name;
         let scalar = &self.scalar;
 
-        let (impl_generics, where_clause) = self.ty.impl_generics(true);
-        let ty = self.ty.ty_tokens();
-        let ty_name = ty.to_string();
-        let trait_ty = self.ty.trait_ty();
+        let generics = self.impl_generics(true);
+        let (impl_generics, _, where_clause) = generics.split_for_impl();
+        let (_, ty_generics, _) = self.trait_generics.split_for_impl();
 
-        let fields_resolvers = self
-            .fields
-            .iter()
-            .map(|f| f.method_resolve_field_async_tokens(scalar, Some(&trait_ty)));
+        let fields_resolvers = self.fields.iter().map(|f| {
+            let name = &f.name;
+            quote! {
+                #name => {
+                    ::juniper::macros::reflection::AsyncField::<
+                        #scalar,
+                        { ::juniper::macros::reflection::fnv1a128(#name) }
+                    >::call(self, info, args, executor)
+                }
+            }
+        });
         let no_field_err =
-            field::Definition::method_resolve_field_err_no_field_tokens(scalar, &ty_name);
+            field::Definition::method_resolve_field_err_no_field_tokens(scalar, trait_name);
 
-        let custom_downcasts = self
-            .implementers
-            .iter()
-            .filter_map(|i| i.method_resolve_into_type_async_tokens(&trait_ty));
-        let regular_downcast = self.ty.method_resolve_into_type_async_tokens();
+        let downcast = self.method_resolve_into_type_async_tokens();
 
         quote! {
             #[allow(deprecated, non_snake_case)]
             #[automatically_derived]
-            impl#impl_generics ::juniper::GraphQLValueAsync<#scalar> for #ty #where_clause
+            impl#impl_generics ::juniper::GraphQLValueAsync<#scalar> for #ty#ty_generics
+                #where_clause
             {
                 fn resolve_field_async<'b>(
                     &'b self,
@@ -709,381 +713,410 @@ impl Definition {
                     _: Option<&'b [::juniper::Selection<'b, #scalar>]>,
                     executor: &'b ::juniper::Executor<'b, 'b, Self::Context, #scalar>
                 ) -> ::juniper::BoxFuture<'b, ::juniper::ExecutionResult<#scalar>> {
-                    #( #custom_downcasts )*
-                    #regular_downcast
+                    #downcast
                 }
             }
         }
     }
 
-    /// Returns generated code implementing [`BaseType`], [`BaseSubTypes`] and
-    /// [`WrappedType`] traits for this [GraphQL interface][1].
+    /// Returns generated code implementing [`BaseType`], [`BaseSubTypes`],
+    /// [`WrappedType`] and [`Fields`] traits for this [GraphQL interface][1].
     ///
-    /// [`GraphQLValueAsync`]: juniper::GraphQLValueAsync
+    /// [`BaseSubTypes`]: juniper::macros::reflection::BaseSubTypes
+    /// [`BaseType`]: juniper::macros::reflection::BaseType
+    /// [`Fields`]: juniper::macros::reflection::Fields
+    /// [`WrappedType`]: juniper::macros::reflection::WrappedType
     /// [1]: https://spec.graphql.org/June2018/#sec-Interfaces
     #[must_use]
-    pub(crate) fn impl_traits_for_reflection_tokens(&self) -> TokenStream {
+    fn impl_traits_for_reflection_tokens(&self) -> TokenStream {
+        let ty = &self.enum_alias_ident;
+        let implementers = &self.implementers;
         let scalar = &self.scalar;
         let name = &self.name;
-        let (impl_generics, where_clause) = self.ty.impl_generics(false);
-        let ty = self.ty.ty_tokens();
-        let implementors = self.implementers.iter().map(|i| &i.ty);
+        let fields = self.fields.iter().map(|f| &f.name);
+
+        let generics = self.impl_generics(false);
+        let (impl_generics, _, where_clause) = generics.split_for_impl();
+        let (_, ty_generics, _) = self.trait_generics.split_for_impl();
 
         quote! {
-            impl #impl_generics ::juniper::macros::reflection::BaseType<#scalar> for #ty
+            #[automatically_derived]
+            impl#impl_generics ::juniper::macros::reflection::BaseType<#scalar>
+                for #ty#ty_generics
                 #where_clause
             {
                 const NAME: ::juniper::macros::reflection::Type = #name;
             }
 
-            impl #impl_generics ::juniper::macros::reflection::BaseSubTypes<#scalar> for #ty
+            #[automatically_derived]
+            impl#impl_generics ::juniper::macros::reflection::BaseSubTypes<#scalar>
+                for #ty#ty_generics
                 #where_clause
             {
                 const NAMES: ::juniper::macros::reflection::Types = &[
                     <Self as ::juniper::macros::reflection::BaseType<#scalar>>::NAME,
-                    #(<#implementors as ::juniper::macros::reflection::BaseType<#scalar>>::NAME),*
+                    #(<#implementers as ::juniper::macros::reflection::BaseType<#scalar>>::NAME),*
                 ];
             }
 
-            impl #impl_generics ::juniper::macros::reflection::WrappedType<#scalar> for #ty
+            #[automatically_derived]
+            impl#impl_generics ::juniper::macros::reflection::WrappedType<#scalar>
+                for #ty#ty_generics
                 #where_clause
             {
                 const VALUE: ::juniper::macros::reflection::WrappedValue = 1;
             }
+
+            #[automatically_derived]
+            impl#impl_generics ::juniper::macros::reflection::Fields<#scalar>
+                for #ty#ty_generics
+                #where_clause
+            {
+                const NAMES: ::juniper::macros::reflection::Names = &[#(#fields),*];
+            }
         }
     }
-}
 
-/// Representation of custom downcast into an [`Implementer`] from a
-/// [GraphQL interface][1] type for code generation.
-///
-/// [1]: https://spec.graphql.org/June2018/#sec-Interfaces
-#[derive(Clone, Debug)]
-enum ImplementerDowncast {
-    /// Downcast is performed via a method of trait describing a
+    /// Returns generated code implementing [`FieldMeta`] for each field of this
     /// [GraphQL interface][1].
     ///
+    /// [`FieldMeta`]: juniper::macros::reflection::FieldMeta
     /// [1]: https://spec.graphql.org/June2018/#sec-Interfaces
-    Method {
-        /// Name of trait method which performs this [`ImplementerDowncast`].
-        name: syn::Ident,
+    fn impl_field_meta_tokens(&self) -> TokenStream {
+        let ty = &self.enum_alias_ident;
+        let context = &self.context;
+        let scalar = &self.scalar;
 
-        /// Indicator whether the trait method accepts a [`Context`] as its
-        /// second argument.
-        ///
-        /// [`Context`]: juniper::Context
-        with_context: bool,
-    },
+        let generics = self.impl_generics(false);
+        let (impl_generics, _, where_clause) = generics.split_for_impl();
+        let (_, ty_generics, _) = self.trait_generics.split_for_impl();
 
-    /// Downcast is performed via some external function.
-    External {
-        /// Path of the external function to be called with.
-        path: syn::ExprPath,
-    },
-}
+        self.fields
+            .iter()
+            .map(|field| {
+                let field_name = &field.name;
+                let mut return_ty = field.ty.clone();
+                generics.replace_type_with_defaults(&mut return_ty);
 
-/// Representation of [GraphQL interface][1] implementer for code generation.
-///
-/// [1]: https://spec.graphql.org/June2018/#sec-Interfaces
-#[derive(Clone, Debug)]
-struct Implementer {
-    /// Rust type that this [GraphQL interface][1] [`Implementer`] is
-    /// represented by.
-    ///
-    /// [1]: https://spec.graphql.org/June2018/#sec-Interfaces
-    ty: syn::Type,
+                let (args_tys, args_names): (Vec<_>, Vec<_>) = field
+                    .arguments
+                    .iter()
+                    .flat_map(|vec| vec.iter())
+                    .filter_map(|arg| match arg {
+                        field::MethodArgument::Regular(arg) => Some((&arg.ty, &arg.name)),
+                        _ => None,
+                    })
+                    .unzip();
 
-    /// Custom [`ImplementerDowncast`] for this [`Implementer`].
-    ///
-    /// If absent, then [`Implementer`] is downcast from an enum variant or a
-    /// trait object.
-    downcast: Option<ImplementerDowncast>,
-
-    /// Rust type of [`Context`] that this [GraphQL interface][1]
-    /// [`Implementer`] requires for downcasting.
-    ///
-    /// It's available only when code generation happens for Rust traits and a
-    /// trait method contains context argument.
-    ///
-    /// [`Context`]: juniper::Context
-    /// [1]: https://spec.graphql.org/June2018/#sec-Interfaces
-    context: Option<syn::Type>,
-
-    /// [`ScalarValue`] parametrization of this [`Implementer`].
-    ///
-    /// [`ScalarValue`]: juniper::ScalarValue
-    scalar: scalar::Type,
-}
-
-impl Implementer {
-    /// Returns generated code of downcasting this [`Implementer`] via custom
-    /// [`ImplementerDowncast`].
-    ///
-    /// Returns [`None`] if there is no custom [`Implementer::downcast`].
-    #[must_use]
-    fn downcast_call_tokens(
-        &self,
-        trait_ty: &syn::Type,
-        ctx: Option<syn::Expr>,
-    ) -> Option<TokenStream> {
-        let ctx = ctx.unwrap_or_else(|| parse_quote! { executor.context() });
-        let mut ctx_arg = Some(quote! { , ::juniper::FromContext::from(#ctx) });
-
-        let fn_path = match self.downcast.as_ref()? {
-            ImplementerDowncast::Method { name, with_context } => {
-                if !with_context {
-                    ctx_arg = None;
+                quote! {
+                    #[allow(non_snake_case)]
+                    impl#impl_generics ::juniper::macros::reflection::FieldMeta<
+                        #scalar,
+                        { ::juniper::macros::reflection::fnv1a128(#field_name) }
+                    > for #ty#ty_generics #where_clause {
+                        type Context = #context;
+                        type TypeInfo = ();
+                        const TYPE: ::juniper::macros::reflection::Type =
+                            <#return_ty as ::juniper::macros::reflection::BaseType<#scalar>>::NAME;
+                        const SUB_TYPES: ::juniper::macros::reflection::Types =
+                            <#return_ty as ::juniper::macros::reflection::BaseSubTypes<#scalar>>::NAMES;
+                        const WRAPPED_VALUE: ::juniper::macros::reflection::WrappedValue =
+                            <#return_ty as ::juniper::macros::reflection::WrappedType<#scalar>>::VALUE;
+                        const ARGUMENTS: &'static [(
+                            ::juniper::macros::reflection::Name,
+                            ::juniper::macros::reflection::Type,
+                            ::juniper::macros::reflection::WrappedValue,
+                        )] = &[#((
+                            #args_names,
+                            <#args_tys as ::juniper::macros::reflection::BaseType<#scalar>>::NAME,
+                            <#args_tys as ::juniper::macros::reflection::WrappedType<#scalar>>::VALUE,
+                        )),*];
+                    }
                 }
-                quote! { <Self as #trait_ty>::#name }
-            }
-            ImplementerDowncast::External { path } => {
-                quote! { #path }
-            }
-        };
-
-        Some(quote! {
-            #fn_path(self #ctx_arg)
-        })
+            })
+            .collect()
     }
 
-    /// Returns generated code for the [`GraphQLValue::concrete_type_name`]
-    /// method, which returns name of the GraphQL type represented by this
-    /// [`Implementer`].
+    /// Returns generated code implementing [`Field`] trait for each field of
+    /// this [GraphQL interface][1].
     ///
-    /// Returns [`None`] if there is no custom [`Implementer::downcast`].
-    ///
-    /// [`GraphQLValue::concrete_type_name`]: juniper::GraphQLValue::concrete_type_name
-    #[must_use]
-    fn method_concrete_type_name_tokens(&self, trait_ty: &syn::Type) -> Option<TokenStream> {
-        self.downcast.as_ref()?;
-
-        let ty = &self.ty;
-        let scalar = &self.scalar;
-
-        let downcast = self.downcast_call_tokens(trait_ty, Some(parse_quote! { context }));
-
-        // Doing this may be quite an expensive, because resolving may contain some heavy
-        // computation, so we're preforming it twice. Unfortunately, we have no other options here,
-        // until the `juniper::GraphQLType` itself will allow to do it in some cleverer way.
-        Some(quote! {
-            if (#downcast as ::std::option::Option<&#ty>).is_some() {
-                return <#ty as ::juniper::GraphQLType<#scalar>>::name(info).unwrap().to_string();
-            }
-        })
-    }
-
-    /// Returns generated code for the [`GraphQLValue::resolve_into_type`][0]
-    /// method, which downcasts the [GraphQL interface][1] type into this
-    /// [`Implementer`] synchronously.
-    ///
-    /// Returns [`None`] if there is no custom [`Implementer::downcast`].
-    ///
-    /// [0]: juniper::GraphQLValue::resolve_into_type
+    /// [`Field`]: juniper::macros::reflection::Field
     /// [1]: https://spec.graphql.org/June2018/#sec-Interfaces
-    #[must_use]
-    fn method_resolve_into_type_tokens(&self, trait_ty: &syn::Type) -> Option<TokenStream> {
-        self.downcast.as_ref()?;
+    fn impl_field_tokens(&self) -> TokenStream {
+        let ty = &self.enum_alias_ident;
+        let scalar = &self.scalar;
+        let const_scalar = self.scalar.default_ty();
 
-        let ty = &self.ty;
-        let ty_name = ty.to_token_stream().to_string();
+        let impl_tys = self.implementers.iter().collect::<Vec<_>>();
+        let impl_idents = self
+            .implementers
+            .iter()
+            .filter_map(|ty| ty.path.segments.last().map(|seg| &seg.ident))
+            .collect::<Vec<_>>();
+
+        let generics = self.impl_generics(false);
+        let (impl_generics, _, where_clause) = generics.split_for_impl();
+        let (_, ty_generics, _) = self.trait_generics.split_for_impl();
+
+        self.fields
+            .iter()
+            .filter_map(|field| {
+                if field.is_async {
+                    return None;
+                }
+
+                let field_name = &field.name;
+                let mut return_ty = field.ty.clone();
+                generics.replace_type_with_defaults(&mut return_ty);
+
+                let const_ty_generics = self.const_trait_generics();
+
+                let unreachable_arm = (self.implementers.is_empty()
+                    || !self.trait_generics.params.is_empty())
+                .then(|| {
+                    quote! { _ => unreachable!() }
+                });
+
+                Some(quote! {
+                    #[allow(non_snake_case)]
+                    impl#impl_generics ::juniper::macros::reflection::Field<
+                        #scalar,
+                        { ::juniper::macros::reflection::fnv1a128(#field_name) }
+                    > for #ty#ty_generics #where_clause {
+                        fn call(
+                            &self,
+                            info: &Self::TypeInfo,
+                            args: &::juniper::Arguments<#scalar>,
+                            executor: &::juniper::Executor<Self::Context, #scalar>,
+                        ) -> ::juniper::ExecutionResult<#scalar> {
+                            match self {
+                                #(#ty::#impl_idents(v) => {
+                                    ::juniper::assert_field!(
+                                        #ty#const_ty_generics,
+                                        #impl_tys,
+                                        #const_scalar,
+                                        #field_name,
+                                    );
+
+                                    <_ as ::juniper::macros::reflection::Field<
+                                        #scalar,
+                                        { ::juniper::macros::reflection::fnv1a128(#field_name) },
+                                    >>::call(v, info, args, executor)
+                                })*
+                                #unreachable_arm
+                            }
+                        }
+                    }
+                })
+            })
+            .collect()
+    }
+
+    /// Returns generated code implementing [`AsyncField`] trait for each field
+    /// of this [GraphQL interface][1].
+    ///
+    /// [`AsyncField`]: juniper::macros::reflection::AsyncField
+    /// [1]: https://spec.graphql.org/June2018/#sec-Interfaces
+    fn impl_async_field_tokens(&self) -> TokenStream {
+        let ty = &self.enum_alias_ident;
+        let scalar = &self.scalar;
+        let const_scalar = self.scalar.default_ty();
+
+        let impl_tys = self.implementers.iter().collect::<Vec<_>>();
+        let impl_idents = self
+            .implementers
+            .iter()
+            .filter_map(|ty| ty.path.segments.last().map(|seg| &seg.ident))
+            .collect::<Vec<_>>();
+
+        let generics = self.impl_generics(true);
+        let (impl_generics, _, where_clause) = generics.split_for_impl();
+        let (_, ty_generics, _) = self.trait_generics.split_for_impl();
+
+        self.fields
+            .iter()
+            .map(|field| {
+                let field_name = &field.name;
+                let mut return_ty = field.ty.clone();
+                generics.replace_type_with_defaults(&mut return_ty);
+
+                let const_ty_generics = self.const_trait_generics();
+
+                let unreachable_arm = (self.implementers.is_empty()
+                    || !self.trait_generics.params.is_empty())
+                .then(|| {
+                    quote! { _ => unreachable!() }
+                });
+
+                quote! {
+                    #[allow(non_snake_case)]
+                    impl#impl_generics ::juniper::macros::reflection::AsyncField<
+                        #scalar,
+                        { ::juniper::macros::reflection::fnv1a128(#field_name) }
+                    > for #ty#ty_generics #where_clause {
+                        fn call<'b>(
+                            &'b self,
+                            info: &'b Self::TypeInfo,
+                            args: &'b ::juniper::Arguments<#scalar>,
+                            executor: &'b ::juniper::Executor<Self::Context, #scalar>,
+                        ) -> ::juniper::BoxFuture<'b, ::juniper::ExecutionResult<#scalar>> {
+                            match self {
+                                #(#ty::#impl_idents(v) => {
+                                    ::juniper::assert_field!(
+                                        #ty#const_ty_generics,
+                                        #impl_tys,
+                                        #const_scalar,
+                                        #field_name,
+                                    );
+
+                                    <_ as ::juniper::macros::reflection::AsyncField<
+                                        #scalar,
+                                        { ::juniper::macros::reflection::fnv1a128(#field_name) },
+                                    >>::call(v, info, args, executor)
+                                })*
+                                #unreachable_arm
+                            }
+                        }
+                    }
+                }
+            })
+            .collect()
+    }
+
+    /// Returns generated code for the [`GraphQLValue::concrete_type_name`][0]
+    /// method, which returns name of the underlying [`Implementer`] GraphQL
+    /// type contained in this [`EnumType`].
+    ///
+    /// [0]: juniper::GraphQLValue::concrete_type_name
+    #[must_use]
+    fn method_concrete_type_name_tokens(&self) -> TokenStream {
         let scalar = &self.scalar;
 
-        let downcast = self.downcast_call_tokens(trait_ty, None);
+        let match_arms = self
+            .implementers
+            .iter()
+            .filter_map(|ty| ty.path.segments.last().map(|seg| (&seg.ident, ty)))
+            .map(|(ident, ty)| {
+                quote! {
+                    Self::#ident(v) => <
+                        #ty as ::juniper::GraphQLValue<#scalar>
+                    >::concrete_type_name(v, context, info),
+                }
+            });
 
-        let resolving_code = gen::sync_resolving_code();
+        let non_exhaustive_match_arm =
+            (!self.trait_generics.params.is_empty() || self.implementers.is_empty()).then(|| {
+                quote! { _ => unreachable!(), }
+            });
 
-        Some(quote! {
-            if type_name == <#ty as ::juniper::GraphQLType<#scalar>>::name(info)
-                .ok_or_else(|| ::juniper::macros::helper::err_unnamed_type(#ty_name))?
-            {
-                let res = #downcast;
-                return #resolving_code;
+        quote! {
+            match self {
+                #( #match_arms )*
+                #non_exhaustive_match_arm
             }
-        })
+        }
     }
 
     /// Returns generated code for the
     /// [`GraphQLValueAsync::resolve_into_type_async`][0] method, which
-    /// downcasts the [GraphQL interface][1] type into this [`Implementer`]
+    /// downcasts this [`EnumType`] into its underlying [`Implementer`] type
     /// asynchronously.
     ///
-    /// Returns [`None`] if there is no custom [`Implementer::downcast`].
-    ///
     /// [0]: juniper::GraphQLValueAsync::resolve_into_type_async
-    /// [1]: https://spec.graphql.org/June2018/#sec-Interfaces
     #[must_use]
-    fn method_resolve_into_type_async_tokens(&self, trait_ty: &syn::Type) -> Option<TokenStream> {
-        self.downcast.as_ref()?;
-
-        let ty = &self.ty;
-        let ty_name = ty.to_token_stream().to_string();
-        let scalar = &self.scalar;
-
-        let downcast = self.downcast_call_tokens(trait_ty, None);
-
+    fn method_resolve_into_type_async_tokens(&self) -> TokenStream {
         let resolving_code = gen::async_resolving_code(None);
 
-        Some(quote! {
-            match <#ty as ::juniper::GraphQLType<#scalar>>::name(info) {
-                Some(name) => {
-                    if type_name == name {
-                        let fut = ::juniper::futures::future::ready(#downcast);
-                        return #resolving_code;
+        let match_arms = self.implementers.iter().filter_map(|ty| {
+            ty.path.segments.last().map(|ident| {
+                quote! {
+                    Self::#ident(v) => {
+                        let fut = ::juniper::futures::future::ready(v);
+                        #resolving_code
                     }
                 }
-                None => return ::juniper::macros::helper::err_unnamed_type_fut(#ty_name),
+            })
+        });
+        let non_exhaustive_match_arm =
+            (!self.trait_generics.params.is_empty() || self.implementers.is_empty()).then(|| {
+                quote! { _ => unreachable!(), }
+            });
+
+        quote! {
+            match self {
+                #( #match_arms )*
+                #non_exhaustive_match_arm
             }
-        })
-    }
-}
-
-/// Representation of Rust enum implementing [GraphQL interface][1] type for
-/// code generation.
-///
-/// [1]: https://spec.graphql.org/June2018/#sec-Interfaces
-struct EnumType {
-    /// Name of this [`EnumType`] to generate it with.
-    ident: syn::Ident,
-
-    /// [`syn::Visibility`] of this [`EnumType`] to generate it with.
-    visibility: syn::Visibility,
-
-    /// Rust types of all [GraphQL interface][1] implements to represent
-    /// variants of this [`EnumType`].
-    ///
-    /// [1]: https://spec.graphql.org/June2018/#sec-Interfaces
-    variants: Vec<syn::Type>,
-
-    /// Name of the trait describing the [GraphQL interface][1] represented by
-    /// this [`EnumType`].
-    ///
-    /// [1]: https://spec.graphql.org/June2018/#sec-Interfaces
-    trait_ident: syn::Ident,
-
-    /// [`syn::Generics`] of the trait describing the [GraphQL interface][1]
-    /// represented by this [`EnumType`].
-    ///
-    /// [1]: https://spec.graphql.org/June2018/#sec-Interfaces
-    trait_generics: syn::Generics,
-
-    /// Associated types of the trait describing the [GraphQL interface][1]
-    /// represented by this [`EnumType`].
-    trait_types: Vec<(syn::Ident, syn::Generics)>,
-
-    /// Associated constants of the trait describing the [GraphQL interface][1]
-    /// represented by this [`EnumType`].
-    ///
-    /// [1]: https://spec.graphql.org/June2018/#sec-Interfaces
-    trait_consts: Vec<(syn::Ident, syn::Type)>,
-
-    /// Methods of the trait describing the [GraphQL interface][1] represented
-    /// by this [`EnumType`].
-    ///
-    /// [1]: https://spec.graphql.org/June2018/#sec-Interfaces
-    trait_methods: Vec<syn::Signature>,
-
-    /// [`ScalarValue`] parametrization to generate [`GraphQLType`]
-    /// implementation with for this [`EnumType`].
-    ///
-    /// [`GraphQLType`]: juniper::GraphQLType
-    /// [`ScalarValue`]: juniper::ScalarValue
-    scalar: scalar::Type,
-}
-
-impl ToTokens for EnumType {
-    fn to_tokens(&self, into: &mut TokenStream) {
-        self.type_definition_tokens().to_tokens(into);
-        into.append_all(self.impl_from_tokens());
-        self.impl_trait_tokens().to_tokens(into);
-    }
-}
-
-impl EnumType {
-    /// Constructs a new [`EnumType`] out of the given parameters.
-    #[must_use]
-    fn new(
-        r#trait: &syn::ItemTrait,
-        meta: &TraitAttr,
-        implers: &[Implementer],
-        scalar: scalar::Type,
-    ) -> Self {
-        Self {
-            ident: meta
-                .r#enum
-                .as_ref()
-                .map(SpanContainer::as_ref)
-                .cloned()
-                .unwrap_or_else(|| format_ident!("{}Value", r#trait.ident)),
-            visibility: r#trait.vis.clone(),
-            variants: implers.iter().map(|impler| impler.ty.clone()).collect(),
-            trait_ident: r#trait.ident.clone(),
-            trait_generics: r#trait.generics.clone(),
-            trait_types: r#trait
-                .items
-                .iter()
-                .filter_map(|i| {
-                    if let syn::TraitItem::Type(ty) = i {
-                        Some((ty.ident.clone(), ty.generics.clone()))
-                    } else {
-                        None
-                    }
-                })
-                .collect(),
-            trait_consts: r#trait
-                .items
-                .iter()
-                .filter_map(|i| {
-                    if let syn::TraitItem::Const(cnst) = i {
-                        Some((cnst.ident.clone(), cnst.ty.clone()))
-                    } else {
-                        None
-                    }
-                })
-                .collect(),
-            trait_methods: r#trait
-                .items
-                .iter()
-                .filter_map(|i| {
-                    if let syn::TraitItem::Method(m) = i {
-                        Some(m.sig.clone())
-                    } else {
-                        None
-                    }
-                })
-                .collect(),
-            scalar,
         }
     }
 
-    /// Returns name of a single variant of this [`EnumType`] by the given
-    /// underlying [`syn::Type`] of the variant.
-    #[must_use]
-    fn variant_ident(ty: &syn::Type) -> &syn::Ident {
-        if let syn::Type::Path(p) = ty {
-            &p.path.segments.last().unwrap().ident
-        } else {
-            unreachable!("GraphQL object has unexpected type `{}`", quote! { #ty })
-        }
-    }
-
-    /// Indicates whether this [`EnumType`] has non-exhaustive phantom variant
-    /// to hold type parameters.
-    #[must_use]
-    fn has_phantom_variant(&self) -> bool {
-        !self.trait_generics.params.is_empty()
-    }
-
-    /// Returns generate code for dispatching non-exhaustive phantom variant of
-    /// this [`EnumType`] in `match` expressions.
+    /// Returns generated code for the [`GraphQLValue::resolve_into_type`][0]
+    /// method, which downcasts this [`EnumType`] into its underlying
+    /// [`Implementer`] type synchronously.
     ///
-    /// Returns [`None`] if this [`EnumType`] is exhaustive.
+    /// [0]: juniper::GraphQLValue::resolve_into_type
     #[must_use]
-    fn non_exhaustive_match_arm_tokens(&self) -> Option<TokenStream> {
-        if self.has_phantom_variant() || self.variants.is_empty() {
-            Some(quote! { _ => unreachable!(), })
-        } else {
-            None
+    fn method_resolve_into_type_tokens(&self) -> TokenStream {
+        let resolving_code = gen::sync_resolving_code();
+
+        let match_arms = self.implementers.iter().filter_map(|ty| {
+            ty.path.segments.last().map(|ident| {
+                quote! {
+                    Self::#ident(res) => #resolving_code,
+                }
+            })
+        });
+
+        let non_exhaustive_match_arm =
+            (!self.trait_generics.params.is_empty() || self.implementers.is_empty()).then(|| {
+                quote! { _ => unreachable!(), }
+            });
+
+        quote! {
+            match self {
+                #( #match_arms )*
+                #non_exhaustive_match_arm
+            }
         }
+    }
+
+    /// Returns trait generics replaced with default values for usage in `const`
+    /// context.
+    #[must_use]
+    fn const_trait_generics(&self) -> syn::PathArguments {
+        struct GenericsForConst(syn::AngleBracketedGenericArguments);
+
+        impl Visit<'_> for GenericsForConst {
+            fn visit_generic_param(&mut self, param: &syn::GenericParam) {
+                let arg = match param {
+                    syn::GenericParam::Lifetime(_) => parse_quote!( 'static ),
+                    syn::GenericParam::Type(ty) => {
+                        if ty.default.is_none() {
+                            parse_quote!(::juniper::DefaultScalarValue)
+                        } else {
+                            return;
+                        }
+                    }
+                    syn::GenericParam::Const(c) => {
+                        if c.default.is_none() {
+                            // This hack works because only `min_const_generics`
+                            // are enabled for now.
+                            // TODO: replace this once full `const_generics` are
+                            //       available.
+                            //       Maybe with `<_ as Default>::default()`?
+                            parse_quote!({ 0_u8 as _ })
+                        } else {
+                            return;
+                        }
+                    }
+                };
+                self.0.args.push(arg)
+            }
+        }
+
+        let mut visitor = GenericsForConst(parse_quote!( <> ));
+        visitor.visit_generics(&self.trait_generics);
+        syn::PathArguments::AngleBracketed(visitor.0)
     }
 
     /// Returns prepared [`syn::Generics`] for [`GraphQLType`] trait (and
@@ -1123,7 +1156,7 @@ impl EnumType {
                 }
 
                 let lifetimes = generics.lifetimes().map(|lt| &lt.lifetime);
-                let ty = &self.ident;
+                let ty = &self.enum_alias_ident;
                 let (_, ty_generics, _) = generics.split_for_impl();
 
                 quote! { for<#( #lifetimes ),*> #ty#ty_generics }
@@ -1146,635 +1179,11 @@ impl EnumType {
         generics
     }
 
-    /// Returns full type signature of the original trait describing the
-    /// [GraphQL interface][1] for this [`EnumType`].
-    ///
-    /// [1]: https://spec.graphql.org/June2018/#sec-Interfaces
+    /// Indicates whether this [`EnumType`] has non-exhaustive phantom variant
+    /// to hold type parameters.
     #[must_use]
-    fn trait_ty(&self) -> syn::Type {
-        let ty = &self.trait_ident;
-        let (_, generics, _) = self.trait_generics.split_for_impl();
-
-        parse_quote! { #ty#generics }
-    }
-
-    /// Returns generated code of the full type signature of this [`EnumType`].
-    #[must_use]
-    fn ty_tokens(&self) -> TokenStream {
-        let ty = &self.ident;
-        let (_, generics, _) = self.trait_generics.split_for_impl();
-
-        quote! { #ty#generics }
-    }
-
-    /// Returns generate code of the Rust type definitions of this [`EnumType`].
-    ///
-    /// If the [`EnumType::trait_generics`] are not empty, then they are
-    /// contained in the generated enum too.
-    #[must_use]
-    fn type_definition_tokens(&self) -> TokenStream {
-        let enum_ty = &self.ident;
-        let generics = &self.trait_generics;
-        let vis = &self.visibility;
-
-        let doc = format!(
-            "Type implementing [GraphQL interface][1] represented by `{}` trait.\
-             \n\n\
-             [1]: https://spec.graphql.org/June2018/#sec-Interfaces",
-            self.trait_ident,
-        );
-
-        let variants = self.variants.iter().map(|ty| {
-            let variant = Self::variant_ident(ty);
-            let doc = format!(
-                "`{}` implementer of this GraphQL interface.",
-                quote! { #ty },
-            );
-
-            quote! {
-                #[doc = #doc]
-                #variant(#ty),
-            }
-        });
-
-        let phantom_variant = if self.has_phantom_variant() {
-            let ty_params = generics.params.iter().map(|p| {
-                let ty = match p {
-                    syn::GenericParam::Type(ty) => {
-                        let ident = &ty.ident;
-                        quote! { #ident }
-                    }
-                    syn::GenericParam::Lifetime(lt) => {
-                        let lifetime = &lt.lifetime;
-                        quote! { &#lifetime () }
-                    }
-                    syn::GenericParam::Const(_) => unimplemented!(),
-                };
-                quote! {
-                    ::std::marker::PhantomData<::std::sync::atomic::AtomicPtr<Box<#ty>>>
-                }
-            });
-
-            Some(quote! {
-                #[doc(hidden)]
-                __Phantom(#( #ty_params ),*),
-            })
-        } else {
-            None
-        };
-
-        quote! {
-            #[automatically_derived]
-            #[doc = #doc]
-            #vis enum #enum_ty#generics {
-                #( #variants )*
-                #phantom_variant
-            }
-        }
-    }
-
-    /// Returns generated code implementing [`From`] trait for this [`EnumType`]
-    /// from its [`EnumType::variants`].
-    fn impl_from_tokens(&self) -> impl Iterator<Item = TokenStream> + '_ {
-        let enum_ty = &self.ident;
-        let (impl_generics, generics, where_clause) = self.trait_generics.split_for_impl();
-
-        self.variants.iter().map(move |ty| {
-            let variant = Self::variant_ident(ty);
-
-            quote! {
-                #[automatically_derived]
-                impl#impl_generics From<#ty> for #enum_ty#generics #where_clause {
-                    fn from(v: #ty) -> Self {
-                        Self::#variant(v)
-                    }
-                }
-            }
-        })
-    }
-
-    /// Returns generated code implementing the original trait describing the
-    /// [GraphQL interface][1] for this [`EnumType`].
-    ///
-    /// [1]: https://spec.graphql.org/June2018/#sec-Interfaces
-    #[must_use]
-    fn impl_trait_tokens(&self) -> TokenStream {
-        let enum_ty = &self.ident;
-
-        let trait_ident = &self.trait_ident;
-        let (impl_generics, generics, where_clause) = self.trait_generics.split_for_impl();
-
-        let var_ty = self.variants.first();
-
-        let assoc_types = self.trait_types.iter().map(|(ty, ty_gen)| {
-            quote! {
-                type #ty#ty_gen = <#var_ty as #trait_ident#generics>::#ty#ty_gen;
-            }
-        });
-
-        let assoc_consts = self.trait_consts.iter().map(|(ident, ty)| {
-            quote! {
-                const #ident: #ty = <#var_ty as #trait_ident#generics>::#ident;
-            }
-        });
-
-        let methods = self.trait_methods.iter().map(|sig| {
-            let method = &sig.ident;
-
-            let mut sig = sig.clone();
-            let mut args = vec![];
-            for (n, arg) in sig.inputs.iter_mut().enumerate() {
-                match arg {
-                    syn::FnArg::Receiver(_) => {}
-                    syn::FnArg::Typed(a) => {
-                        if !matches!(&*a.pat, syn::Pat::Ident(_)) {
-                            let ident = format_ident!("__arg{}", n);
-                            a.pat = parse_quote! { #ident };
-                        }
-                        args.push(a.pat.clone());
-                    }
-                }
-            }
-
-            let and_await = if sig.asyncness.is_some() {
-                Some(quote! { .await })
-            } else {
-                None
-            };
-
-            let match_arms = self.variants.iter().map(|ty| {
-                let variant = Self::variant_ident(ty);
-                let args = args.clone();
-
-                quote! {
-                    Self::#variant(v) =>
-                        <#ty as #trait_ident#generics>::#method(v #( , #args )* )#and_await,
-                }
-            });
-            let non_exhaustive_match_arm = self.non_exhaustive_match_arm_tokens();
-
-            quote! {
-                #sig {
-                    match self {
-                        #( #match_arms )*
-                        #non_exhaustive_match_arm
-                    }
-                }
-            }
-        });
-
-        let mut impl_tokens = quote! {
-            #[allow(deprecated)]
-            #[automatically_derived]
-            impl#impl_generics #trait_ident#generics for #enum_ty#generics #where_clause {
-                #( #assoc_types )*
-
-                #( #assoc_consts )*
-
-                #( #methods )*
-            }
-        };
-
-        if self.trait_methods.iter().any(|sig| sig.asyncness.is_some()) {
-            let mut ast: syn::ItemImpl = parse_quote! { #impl_tokens };
-            inject_async_trait(
-                &mut ast.attrs,
-                ast.items.iter_mut().filter_map(|i| {
-                    if let syn::ImplItem::Method(m) = i {
-                        Some(&mut m.sig)
-                    } else {
-                        None
-                    }
-                }),
-                &ast.generics,
-            );
-            impl_tokens = quote! { #ast };
-        }
-
-        impl_tokens
-    }
-
-    /// Returns generated code for the [`GraphQLValue::concrete_type_name`][0]
-    /// method, which returns name of the underlying [`Implementer`] GraphQL
-    /// type contained in this [`EnumType`].
-    ///
-    /// [0]: juniper::GraphQLValue::concrete_type_name
-    #[must_use]
-    fn method_concrete_type_name_tokens(&self) -> TokenStream {
-        let scalar = &self.scalar;
-
-        let match_arms = self.variants.iter().map(|ty| {
-            let variant = Self::variant_ident(ty);
-
-            quote! {
-                Self::#variant(v) => <
-                    #ty as ::juniper::GraphQLValue<#scalar>
-                >::concrete_type_name(v, context, info),
-            }
-        });
-        let non_exhaustive_match_arm = self.non_exhaustive_match_arm_tokens();
-
-        quote! {
-            match self {
-                #( #match_arms )*
-                #non_exhaustive_match_arm
-            }
-        }
-    }
-
-    /// Returns generated code for the [`GraphQLValue::resolve_into_type`][0]
-    /// method, which downcasts this [`EnumType`] into its underlying
-    /// [`Implementer`] type synchronously.
-    ///
-    /// [0]: juniper::GraphQLValue::resolve_into_type
-    #[must_use]
-    fn method_resolve_into_type_tokens(&self) -> TokenStream {
-        let resolving_code = gen::sync_resolving_code();
-
-        let match_arms = self.variants.iter().map(|ty| {
-            let variant = Self::variant_ident(ty);
-
-            quote! {
-                Self::#variant(res) => #resolving_code,
-            }
-        });
-        let non_exhaustive_match_arm = self.non_exhaustive_match_arm_tokens();
-
-        quote! {
-            match self {
-                #( #match_arms )*
-                #non_exhaustive_match_arm
-            }
-        }
-    }
-
-    /// Returns generated code for the
-    /// [`GraphQLValueAsync::resolve_into_type_async`][0] method, which
-    /// downcasts this [`EnumType`] into its underlying [`Implementer`] type
-    /// asynchronously.
-    ///
-    /// [0]: juniper::GraphQLValueAsync::resolve_into_type_async
-    #[must_use]
-    fn method_resolve_into_type_async_tokens(&self) -> TokenStream {
-        let resolving_code = gen::async_resolving_code(None);
-
-        let match_arms = self.variants.iter().map(|ty| {
-            let variant = Self::variant_ident(ty);
-
-            quote! {
-                Self::#variant(v) => {
-                    let fut = ::juniper::futures::future::ready(v);
-                    #resolving_code
-                }
-            }
-        });
-        let non_exhaustive_match_arm = self.non_exhaustive_match_arm_tokens();
-
-        quote! {
-            match self {
-                #( #match_arms )*
-                #non_exhaustive_match_arm
-            }
-        }
-    }
-}
-
-/// Representation of Rust [trait object][2] implementing [GraphQL interface][1]
-/// type for code generation.
-///
-/// [1]: https://spec.graphql.org/June2018/#sec-Interfaces
-/// [2]: https://doc.rust-lang.org/reference/types/trait-object.html
-struct TraitObjectType {
-    /// Name of this [`TraitObjectType`] to generate it with.
-    ident: syn::Ident,
-
-    /// [`syn::Visibility`] of this [`TraitObjectType`] to generate it with.
-    visibility: syn::Visibility,
-
-    /// Name of the trait describing the [GraphQL interface][1] represented by
-    /// this [`TraitObjectType`].
-    ///
-    /// [1]: https://spec.graphql.org/June2018/#sec-Interfaces
-    trait_ident: syn::Ident,
-
-    /// [`syn::Generics`] of the trait describing the [GraphQL interface][1]
-    /// represented by this [`TraitObjectType`].
-    ///
-    /// [1]: https://spec.graphql.org/June2018/#sec-Interfaces
-    trait_generics: syn::Generics,
-
-    /// [`ScalarValue`] parametrization of this [`TraitObjectType`] to generate
-    /// it with.
-    ///
-    /// [`ScalarValue`]: juniper::ScalarValue
-    scalar: scalar::Type,
-
-    /// Rust type of [`Context`] to generate this [`TraitObjectType`] with.
-    ///
-    /// [`Context`]: juniper::Context
-    context: syn::Type,
-}
-
-impl TraitObjectType {
-    /// Constructs a new [`TraitObjectType`] out of the given parameters.
-    #[must_use]
-    fn new(
-        r#trait: &syn::ItemTrait,
-        meta: &TraitAttr,
-        scalar: scalar::Type,
-        context: syn::Type,
-    ) -> Self {
-        Self {
-            ident: meta.r#dyn.as_ref().unwrap().as_ref().clone(),
-            visibility: r#trait.vis.clone(),
-            trait_ident: r#trait.ident.clone(),
-            trait_generics: r#trait.generics.clone(),
-            scalar,
-            context,
-        }
-    }
-
-    /// Returns prepared [`syn::Generics`] for [`GraphQLType`] trait (and
-    /// similar) implementation of this [`TraitObjectType`].
-    ///
-    /// If `for_async` is `true`, then additional predicates are added to suit
-    /// the [`GraphQLAsyncValue`] trait (and similar) requirements.
-    ///
-    /// [`GraphQLAsyncValue`]: juniper::GraphQLAsyncValue
-    /// [`GraphQLType`]: juniper::GraphQLType
-    #[must_use]
-    fn impl_generics(&self, for_async: bool) -> syn::Generics {
-        let mut generics = self.trait_generics.clone();
-
-        generics.params.push(parse_quote! { '__obj });
-
-        let scalar = &self.scalar;
-        if scalar.is_implicit_generic() {
-            generics.params.push(parse_quote! { #scalar });
-        }
-        if scalar.is_generic() {
-            generics
-                .make_where_clause()
-                .predicates
-                .push(parse_quote! { #scalar: ::juniper::ScalarValue });
-        }
-        if let Some(bound) = scalar.bounds() {
-            generics.make_where_clause().predicates.push(bound);
-        }
-
-        if for_async {
-            generics
-                .make_where_clause()
-                .predicates
-                .push(parse_quote! { Self: Sync });
-            if scalar.is_generic() {
-                generics
-                    .make_where_clause()
-                    .predicates
-                    .push(parse_quote! { #scalar: Send + Sync });
-            }
-        }
-
-        generics
-    }
-
-    /// Returns full type signature of the original trait describing the
-    /// [GraphQL interface][1] for this [`TraitObjectType`].
-    ///
-    /// [1]: https://spec.graphql.org/June2018/#sec-Interfaces
-    #[must_use]
-    fn trait_ty(&self) -> syn::Type {
-        let ty = &self.trait_ident;
-
-        let mut generics = self.trait_generics.clone();
-        if !self.scalar.is_explicit_generic() {
-            let scalar = &self.scalar;
-            generics.params.push(parse_quote! { #scalar });
-        }
-        let (_, generics, _) = generics.split_for_impl();
-
-        parse_quote! { #ty#generics }
-    }
-
-    /// Returns generated code of the full type signature of this
-    /// [`TraitObjectType`].
-    #[must_use]
-    fn ty_tokens(&self) -> TokenStream {
-        let ty = &self.trait_ident;
-
-        let mut generics = self.trait_generics.clone();
-        generics.remove_defaults();
-        generics.move_bounds_to_where_clause();
-        if !self.scalar.is_explicit_generic() {
-            let scalar = &self.scalar;
-            generics.params.push(parse_quote! { #scalar });
-        }
-        let ty_params = &generics.params;
-
-        let context = &self.context;
-
-        quote! {
-            dyn #ty<#ty_params, Context = #context, TypeInfo = ()> + '__obj + Send + Sync
-        }
-    }
-
-    /// Returns generated code for the [`GraphQLValue::concrete_type_name`][0]
-    /// method, which returns name of the underlying [`Implementer`] GraphQL
-    /// type contained in this [`TraitObjectType`].
-    ///
-    /// [0]: juniper::GraphQLValue::concrete_type_name
-    #[must_use]
-    fn method_concrete_type_name_tokens(&self) -> TokenStream {
-        quote! {
-            self.as_dyn_graphql_value().concrete_type_name(context, info)
-        }
-    }
-
-    /// Returns generated code for the [`GraphQLValue::resolve_into_type`][0]
-    /// method, which downcasts this [`TraitObjectType`] into its underlying
-    /// [`Implementer`] type synchronously.
-    ///
-    /// [0]: juniper::GraphQLValue::resolve_into_type
-    #[must_use]
-    fn method_resolve_into_type_tokens(&self) -> TokenStream {
-        let resolving_code = gen::sync_resolving_code();
-
-        quote! {
-            let res = self.as_dyn_graphql_value();
-            #resolving_code
-        }
-    }
-
-    /// Returns generated code for the
-    /// [`GraphQLValueAsync::resolve_into_type_async`][0] method, which
-    /// downcasts this [`TraitObjectType`] into its underlying [`Implementer`]
-    /// type asynchronously.
-    ///
-    /// [0]: juniper::GraphQLValueAsync::resolve_into_type_async
-    #[must_use]
-    fn method_resolve_into_type_async_tokens(&self) -> TokenStream {
-        let resolving_code = gen::async_resolving_code(None);
-
-        quote! {
-            let fut = ::juniper::futures::future::ready(self.as_dyn_graphql_value_async());
-            #resolving_code
-        }
-    }
-}
-
-impl ToTokens for TraitObjectType {
-    fn to_tokens(&self, into: &mut TokenStream) {
-        let dyn_ty = &self.ident;
-        let vis = &self.visibility;
-
-        let doc = format!(
-            "Helper alias for the `{}` [trait object][2] implementing [GraphQL interface][1].\
-             \n\n\
-             [1]: https://spec.graphql.org/June2018/#sec-Interfaces\n\
-             [2]: https://doc.rust-lang.org/reference/types/trait-object.html",
-            self.trait_ident,
-        );
-
-        let trait_ident = &self.trait_ident;
-
-        let mut generics = self.trait_generics.clone();
-        if !self.scalar.is_explicit_generic() {
-            let scalar_ty = self.scalar.generic_ty();
-            let default_ty = self.scalar.default_ty();
-            generics
-                .params
-                .push(parse_quote! { #scalar_ty = #default_ty });
-        }
-
-        let (mut ty_params_left, mut ty_params_right) = (None, None);
-        if !generics.params.is_empty() {
-            // We should preserve defaults for left side.
-            generics.move_bounds_to_where_clause();
-            let params = &generics.params;
-            ty_params_left = Some(quote! { , #params });
-
-            generics.remove_defaults();
-            let params = &generics.params;
-            ty_params_right = Some(quote! { #params, });
-        };
-
-        let context = &self.context;
-
-        let dyn_alias = quote! {
-            #[automatically_derived]
-            #[doc = #doc]
-            #vis type #dyn_ty<'a #ty_params_left> =
-                dyn #trait_ident<#ty_params_right Context = #context, TypeInfo = ()> +
-                    'a + Send + Sync;
-        };
-
-        into.append_all(&[dyn_alias]);
-    }
-}
-
-/// Representation of possible Rust types implementing [GraphQL interface][1]
-/// type for code generation.
-///
-/// [1]: https://spec.graphql.org/June2018/#sec-Interfaces
-enum Type {
-    /// [GraphQL interface][1] type implementation as Rust enum.
-    ///
-    /// [1]: https://spec.graphql.org/June2018/#sec-Interfaces
-    Enum(Box<EnumType>),
-
-    /// [GraphQL interface][1] type implementation as Rust [trait object][2].
-    ///
-    /// [1]: https://spec.graphql.org/June2018/#sec-Interfaces
-    /// [2]: https://doc.rust-lang.org/reference/types/trait-object.html
-    TraitObject(Box<TraitObjectType>),
-}
-
-impl ToTokens for Type {
-    fn to_tokens(&self, into: &mut TokenStream) {
-        match self {
-            Self::Enum(e) => e.to_tokens(into),
-            Self::TraitObject(o) => o.to_tokens(into),
-        }
-    }
-}
-
-impl Type {
-    /// Returns prepared [`syn::Generics`] for [`GraphQLType`] trait (and
-    /// similar) implementation of this [`Type`].
-    ///
-    /// If `for_async` is `true`, then additional predicates are added to suit
-    /// the [`GraphQLAsyncValue`] trait (and similar) requirements.
-    ///
-    /// [`GraphQLAsyncValue`]: juniper::GraphQLAsyncValue
-    /// [`GraphQLType`]: juniper::GraphQLType
-    #[must_use]
-    fn impl_generics(&self, for_async: bool) -> (TokenStream, Option<syn::WhereClause>) {
-        let generics = match self {
-            Self::Enum(e) => e.impl_generics(for_async),
-            Self::TraitObject(o) => o.impl_generics(for_async),
-        };
-        let (impl_generics, _, where_clause) = generics.split_for_impl();
-        (quote! { #impl_generics }, where_clause.cloned())
-    }
-
-    /// Returns full type signature of the original trait describing the
-    /// [GraphQL interface][1] for this [`Type`].
-    ///
-    /// [1]: https://spec.graphql.org/June2018/#sec-Interfaces
-    #[must_use]
-    fn trait_ty(&self) -> syn::Type {
-        match self {
-            Self::Enum(e) => e.trait_ty(),
-            Self::TraitObject(o) => o.trait_ty(),
-        }
-    }
-
-    /// Returns generated code of the full type signature of this [`Type`].
-    #[must_use]
-    fn ty_tokens(&self) -> TokenStream {
-        match self {
-            Self::Enum(e) => e.ty_tokens(),
-            Self::TraitObject(o) => o.ty_tokens(),
-        }
-    }
-
-    /// Returns generated code for the [`GraphQLValue::concrete_type_name`][0]
-    /// method, which returns name of the underlying [`Implementer`] GraphQL
-    /// type contained in this [`Type`].
-    ///
-    /// [0]: juniper::GraphQLValue::concrete_type_name
-    #[must_use]
-    fn method_concrete_type_name_tokens(&self) -> TokenStream {
-        match self {
-            Self::Enum(e) => e.method_concrete_type_name_tokens(),
-            Self::TraitObject(o) => o.method_concrete_type_name_tokens(),
-        }
-    }
-
-    /// Returns generated code for the [`GraphQLValue::resolve_into_type`][0]
-    /// method, which downcasts this [`Type`] into its underlying
-    /// [`Implementer`] type synchronously.
-    ///
-    /// [0]: juniper::GraphQLValue::resolve_into_type
-    #[must_use]
-    fn method_resolve_into_type_tokens(&self) -> TokenStream {
-        match self {
-            Self::Enum(e) => e.method_resolve_into_type_tokens(),
-            Self::TraitObject(o) => o.method_resolve_into_type_tokens(),
-        }
-    }
-
-    /// Returns generated code for the
-    /// [`GraphQLValueAsync::resolve_into_type_async`][0] method, which
-    /// downcasts this [`Type`] into its underlying [`Implementer`] type
-    /// asynchronously.
-    ///
-    /// [0]: juniper::GraphQLValueAsync::resolve_into_type_async
-    fn method_resolve_into_type_async_tokens(&self) -> TokenStream {
-        match self {
-            Self::Enum(e) => e.method_resolve_into_type_async_tokens(),
-            Self::TraitObject(o) => o.method_resolve_into_type_async_tokens(),
-        }
+    fn has_phantom_variant(&self) -> bool {
+        !self.trait_generics.params.is_empty()
     }
 }
